requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic-settings==2.6.1
python-dotenv==1.0.1
cryptography>=42.0.0
py-opengauss==1.3.10
